🔨 refacto: extract service from intents
